Got projectiles to shoot straight
