+++ client/public/index.html
@@ -2,10 +2,15 @@
 <html>
     <head>
         <meta charset="utf-8">
+        <link rel="stylesheet" href="styles/styles.css">
     </head>
     <body>
-        <script src="/socket.io/socket.io.js"></script>
-        <script src="//cdn.jsdelivr.net/npm/phaser@3.15.1/dist/phaser.min.js"></script>
-        <script src="js/game.js"></script>
+        <div id="game">
+            <script src="/socket.io/socket.io.js"></script>
+            <script src="//cdn.jsdelivr.net/npm/phaser@3.15.1/dist/phaser.min.js"></script>
+            <script src="js/bullet.js"></script>
+            <script src="js/bullets.js"></script>
+            <script src="js/game.js"></script>
+        </div>
     </body>
 </html>
