+++ client/public/styles/styles.css
@@ -1,0 +1,4 @@
+canvas {
+    display: block;
+    margin: auto;
+}
