+++ client/server.js
@@ -2,6 +2,7 @@ const express = require('express');
 const app = express();
 const server = require('http').Server(app);
 const io = require('socket.io')(server);
+// const auServ = require('socket.io-client')('http://localhost:8081/')
 
 app.use(express.static(__dirname + '/public'));
